updating memory controller file and seeting up front end

+++ backend/controllers/memory.controller.js
@@ -44,12 +44,15 @@ export const createMemory = async (req, res) => {
 
 export const deleteMemory = async (req, res) => {
     const {id} = req.params;
+    if(!mongoose.Types.ObjectId.isValid(id)){
+        return res.status(400).json({success:false, message:"Invalid Memory ID"});
+    }
     //console.log("id:", id);  //print out the memory id
     try {
         await Memory.findByIdAndDelete(id);
         res.status(200).json({ success: true, message: "Memory deleted successfully"});
     } catch (error) {
         console.log("Error in deleting memory:", error.message);
-        res.status(404).json({success:false, message:"Memory not found"});
+        res.status(500).json({success:false, message:"Server Error"});
     }
 }
